backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-middleware==1.5.0
django-filter==2.3.0
django-localflavor==3.0.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.6
python-stdnum==1.14
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
wincertstore==0.2
zipp==3.1.0
